backend/requirements.txt
fastapi
uvicorn[standard]
asyncpg
python-dotenv
google-genai
openai
anthropic

# testing
pytest
pytest-asyncio
httpx
